backend/cbt/requirements.txt
autopep8==1.6.0
certifi==2022.6.15
click==8.1.3
cloudinary==1.29.0
colorama==0.4.5
Flask==2.1.3
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
postgres==4.0
psycopg2-binary==2.9.3
psycopg2-pool==1.1
pycodestyle==2.8.0
python-dotenv==0.20.0
six==1.16.0
SQLAlchemy==1.4.39
toml==0.10.2
urllib3==1.26.11
Werkzeug==2.2.0
